# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for ISHC project.",
        "title": "ISHC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/event_tasks": {
            "get": {
                "description": "Get details of all event tasks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "eventtask"
                ],
                "summary": "Get all event tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventTask"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new event task with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "eventtask"
                ],
                "summary": "Create a new event task",
                "parameters": [
                    {
                        "description": "Event Task payload",
                        "name": "eventTask",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EventTask"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventTask"
                        }
                    }
                }
            }
        },
        "/event_tasks/{id}": {
            "get": {
                "description": "Get details of an event task by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "eventtask"
                ],
                "summary": "Get an event task by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventTask"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an event task with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "eventtask"
                ],
                "summary": "Update an event task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Event Task payload",
                        "name": "eventTask",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EventTask"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventTask"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an event task by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "eventtask"
                ],
                "summary": "Delete an event task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/tasks": {
            "get": {
                "description": "Get details of all tasks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Get all tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Task"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a task with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Update a task",
                "parameters": [
                    {
                        "description": "Task payload",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "Create a new task with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Create a new task",
                "parameters": [
                    {
                        "description": "Task payload",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/tasks/{id}": {
            "delete": {
                "description": "Delete a task by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Delete a task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/tasks/{id}/finish": {
            "put": {
                "description": "Mark a task as finished by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Finish a task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/video_monitors": {
            "get": {
                "description": "Get details of all video monitor records",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videomonitor"
                ],
                "summary": "Get all video monitor records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VideoMonitor"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new video monitor record with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videomonitor"
                ],
                "summary": "Create a new video monitor record",
                "parameters": [
                    {
                        "description": "Video Monitor payload",
                        "name": "video",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VideoMonitor"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/video_monitors/{id}": {
            "delete": {
                "description": "Delete a video monitor record by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videomonitor"
                ],
                "summary": "Delete a video monitor record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Video Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "models.CustomTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
...
        "models.EventTask": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.Task": {
            "type": "object",
            "properties": {
                "camera_id": {
                    "type": "integer"
                },
                "end_time": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "id": {
                    "type": "integer"
                },
                "start_time": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "status": {
                    "type": "string"
                },
                "task_type": {
                    "type": "string"
                },
                "url_string": {
                    "type": "string"
                }
            }
        },
        "models.VideoMonitor": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "owner": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    }
}